+++ lib/shortcuts.ps1
@@ -20,11 +20,9 @@ function create_startmenu_shortcuts($manifest, $dir, $global, $arch) {
 }
 
 function shortcut_folder($global) {
-    $directory = [System.IO.Path]::Combine([Environment]::GetFolderPath('startmenu'), 'Programs', 'Scoop Apps')
-    if($global) {
-        $directory = [System.IO.Path]::Combine([Environment]::GetFolderPath('commonstartmenu'), 'Programs', 'Scoop Apps')
-    }
-    return $(ensure $directory)
+    $directory = 'startmenu'
+    if ($global) { $directory = "common$directory" }
+    return Convert-Path $(ensure [System.IO.Path]::Combine([Environment]::GetFolderPath($directory), 'Programs', 'Scoop Apps'))
 }
 
 function startmenu_shortcut([System.IO.FileInfo] $target, $shortcutName, $arguments, [System.IO.FileInfo]$icon, $global) {
@@ -37,7 +35,7 @@ function startmenu_shortcut([System.IO.FileInfo] $target, $shortcutName, $argume
         return
     }
 
-    $scoop_startmenu_folder = Convert-Path $(shortcut_folder $global)
+    $scoop_startmenu_folder = shortcut_folder $global
     $subdirectory = [System.IO.Path]::GetDirectoryName($shortcutName)
     if ($subdirectory) {
         $subdirectory = ensure $([System.IO.Path]::Combine($scoop_startmenu_folder, $subdirectory))
@@ -62,23 +60,10 @@ function rm_startmenu_shortcuts($manifest, $global, $arch) {
     $shortcuts = @(arch_specific 'shortcuts' $manifest $arch)
     $shortcuts | Where-Object { $_ -ne $null } | ForEach-Object {
         $name = $_.item(1)
-        $shortcut = "$(Convert-Path $(shortcut_folder $global))\$name.lnk"
+        $shortcut = "$(shortcut_folder $global)\$name.lnk"
         write-host "Removing shortcut $(friendly_path $shortcut)"
         if(Test-Path -Path $shortcut) {
              Remove-Item $shortcut
-        }
-        # Before issue 1514 Startmenu shortcut removal
-        #
-        # Shortcuts that should have been installed globally would
-        # have been installed locally up until 27 June 2017.
-        #
-        # TODO: Remove this 'if' block and comment after
-        #       27 June 2018.
-        if($global) {
-            $shortcut = "$(Convert-Path $(shortcut_folder $false))\$name.lnk"
-            if(Test-Path -Path $shortcut) {
-                 Remove-Item $shortcut
-            }
         }
     }
 }